Pass arguments to the application
The application can now process (very softly for now) a
positional argument (the .json file) and an optional argument to
output the .csv file (unused for now).

In case of a non-existing file passed as argument the application
will fail. By default, the optional output file is an empty string
but it might be optimized later.

Two tests are provided to test the app's behavior with an existing
file and with a non-existing file passed as arguments.