Improve json file validity checks
We were only checking if the file existed, but not its
extension. Hence a more robust checking is done to also check
the extension of the file passed as argument.

At some point we will also have to check if the json file passed
as input is indeed from GitHub issues.