Fix output checking tests behavior
To make tests fail, "panic" was used instead of a proper error
being returned. This has been fixed taking advantage of Box.

Also, all the output files created had the same name which caused
issues since tests are run in parallel: the output file for a test
would get removed when an other was deleting it.